+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log
 
-## [0.4.43] - 2019-05-
+## [0.4.43] - 2019-05-16
+### Added
+- The `tag` parameter for `/list`, `/api/list`, and
+  `interview_menu()`.
+- The `DAContext` object and the `use objects` option for the `data`
+  and `data from code` blocks.
 ### Fixed
-- HTML error when `hide standard menu` in use.
+- HTML error when `hide standard menu` is in use.
 
 ## [0.4.42] - 2019-05-11
 ### Changed